fixing missing methods in view class: moveInTopLeftCorner, moveinTopRightCorner, moveInBottomLeftCorner, moveInBottomRightCorner part 2
