+++ .gitignore
@@ -31,3 +31,4 @@ node_modules
 
 # Optional REPL history
 .node_repl_history
+.idea/**
