+++ demo/src/CandyButtonDemoView.js
@@ -3,29 +3,9 @@ import DemoView from './DemoView';
 import Button from 'canvas_ui/Button';
 import {getContext} from 'canvas_ui/core';
 import {BoldSystemFontOfSize} from 'canvas_ui/fonts';
+import CandyButton from './CandyButton';
 
 let CandyButtonHeight = 100;
-
-export default class CandyButton extends Button {
-    constructor(text, color1, color2, textColor = '#fff') {
-        super(text);
-
-        let context = getContext();
-
-        let grad = context.createLinearGradient(0, 0, 0, CandyButtonHeight);
-        grad.addColorStop(0, color1);
-        grad.addColorStop(1, color2);
-        this.backgroundColor = grad;
-
-        this.font = BoldSystemFontOfSize(20);
-        this.textColor = textColor;
-
-        this.borderWidth = 4;
-        this.borderColor = textColor;
-
-        this.cornerRadius = 20;
-    }
-}
 
 export default class CandyButtonDemoView extends DemoView {
     constructor() {
@@ -41,7 +21,7 @@ export default class CandyButtonDemoView extends DemoView {
     layoutSubviews() {
         _.each(this.subviews, subview => {
             subview.size.set(Math.max(100, this.superview.width/2),
-                             CandyButtonHeight);
+                CandyButtonHeight);
         });
 
         this.playNowButton.moveToCenterMiddle();
